Bugfix#171
Changed the __minigames array into a 2Darray that now also to holds a boolean value for each minigame show if the minigame has been completed.

Added a while loop that checks the boolean value to make sure a minigame is not picked twice.

Added two variables to App. isMinigameCompleted & lastMinigame

**More notes on the code

**Make sure that line 54 in GoodGuyGame works for setting the current minigame.