works maybe idk help
Hopefully when the game ends through voteTally() it will return either "draw" or the dict key associated to the player to be killed.

Then in voting game it will get the players name and check their team and display it all accordingly.

Need to add an execute function to wifi so it can call execute player from voting game