execute player in votingGame
Added a function to call executeFunction() from model.py in wifi.py.

endVote() now will return a list containing the playerID and the value from playerID in the players dict.
So that the playerID and the values containing player name and team can all be used in votingGame.py

THIS NEEDS TO BE CAREFULLY REVIEWED - This has many contingencies and can go wrong in many places as i am not 100% about how some stuff will be returned